+++ Communication/Packets/Incoming/Users/GetHabboGroupBadgesEvent.cs
@@ -12,11 +12,9 @@ internal class GetHabboGroupBadgesEvent : IPacketEvent
 
     public Task Parse(GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        if (habbo == null || !habbo.InRoom)
+        if (session.GetHabbo() is not { InRoom: true } habbo || !habbo.TryGetCurrentRoom(out var room))
             return Task.CompletedTask;
-        if (!habbo.TryGetCurrentRoom(out var room))
-            return Task.CompletedTask;
+
         var badges = _groupManager.GetAllBadgesInRoom(room);
         if (badges != null)
         {